LJS-118 updated test cases
