--- packages/ag-grid/.storybook/addons.js
@@ -1,2 +1,0 @@
-import 'storybook-readme/register';
-import '@storybook/addon-storysource/register';
