--- packages/ag-grid/.storybook/webpack.config.js
@@ -1,19 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  module: {
-    rules: [
-      {
-        test: /\.jsx?$/,
-        include: path.resolve('./stories'),
-        loader: require.resolve('@storybook/addon-storysource/loader'),
-        options: { prettierConfig: { parser: 'babel' } },
-        enforce: 'pre'
-      },
-      {
-        test: /\.css$/,
-        use: ['style-loader', 'css-loader']
-      }
-    ]
-  }
-};
