+++ packages/ag-grid/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@latticejs/ag-grid",
-  "version": "1.0.1-beta.3",
+  "version": "1.0.1-beta.4",
   "license": "MIT",
   "module": "dist/esm/index.js",
   "browser": "dist/umd/ag-grid.js",
@@ -10,13 +10,14 @@
   ],
   "scripts": {
     "build:clean": "rimraf dist/**",
-    "build": "npm run build:clean && lattice-scripts build ./src/index.js",
-    "build:dev": "npm run build:clean && lattice-scripts build -e development ./src/index.js",
+    "build": "npm run build:clean && lattice-scripts build ./src/components/index.js",
+    "build:dev": "npm run build:clean && lattice-scripts build -e development ./src/components/index.js",
     "release": "npm run build",
     "prepublish": "npm run release",
-    "storybook": "start-storybook -p 9016 -c .storybook",
+    "storybook": "start-storybook -p 6006",
     "test": "jest --no-cache",
-    "size": "bundlesize"
+    "size": "bundlesize",
+    "build-storybook": "build-storybook"
   },
   "peerDependencies": {
     "@material-ui/core": "^4.8.2",
@@ -27,6 +28,7 @@
   "dependencies": {
     "ag-grid-community": "^19.1.4",
     "ag-grid-react": "^19.1.2",
+    "axios": "^0.20.0",
     "typeface-roboto": "^0.0.54"
   },
   "devDependencies": {
@@ -39,28 +41,36 @@
     "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-react": "^7.10.1",
+    "@latticejs/lattice-scripts": "^1.0.1-beta.2",
     "@material-ui/core": "^4.8.2",
     "@material-ui/icons": "^4.5.1",
-    "@latticejs/lattice-scripts": "^1.0.1-beta.2",
-    "@storybook/addon-storyshots": "^5.3.19",
-    "@storybook/addon-storysource": "^5.3.19",
-    "@storybook/react": "^5.3.19",
+    "@storybook/addon-actions": "^6.0.26",
+    "@storybook/addon-controls": "^6.0.27",
+    "@storybook/addon-docs": "^6.0.26",
+    "@storybook/addon-essentials": "^6.0.26",
+    "@storybook/addon-links": "^6.0.26",
+    "@storybook/addon-storyshots": "^6.0.26",
+    "@storybook/addon-storysource": "^6.0.26",
+    "@storybook/react": "^6.0.26",
+    "babel-loader": "^8.1.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.1",
     "jest": "^23.6.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-is": "^16.13.1",
+    "react-scripts": "^1.1.4",
     "react-test-renderer": "^16.13.1",
     "rimraf": "^2.7.0",
-    "storybook-readme": "^4.0.5",
-    "react-scripts": "^1.1.4"
+    "storybook-readme": "^5.0.8"
   },
   "jest": {
     "setupTestFrameworkScriptFile": "./setupTest.js",
     "testURL": "http://localhost",
     "transform": {
       "^.+\\.js?$": "babel-jest",
-      "^.+\\.md?$": "markdown-loader-jest"
+      "^.+\\.md?$": "markdown-loader-jest",
+      "^.+\\.mdx$": "@storybook/addon-docs/jest-transform-mdx"
     },
     "testPathIgnorePatterns": [
       "<rootDir>/images/",
@@ -84,6 +94,12 @@
   },
   "description": "Standalone Ag Grid Component of Lattice blended with Material UI",
   "main": "dist/cjs/index.js",
-  "keywords": ["Lattice", "Ag-Grid", "Ag-Grid Lattice", "Lattice Ag-Grid", "Grid"],
+  "keywords": [
+    "Lattice",
+    "Ag-Grid",
+    "Ag-Grid Lattice",
+    "Lattice Ag-Grid",
+    "Grid"
+  ],
   "author": ""
 }
